refactor: extract Semaphore utility
